Abstract profile body
