+++ need2give/lib/screens/main_pages_navbar/button_navbar.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:need2give/constants/global.dart';
 import 'package:need2give/screens/donation_center/dashboard.dart';
+import 'package:need2give/screens/donation_center/profile.dart';
 import 'package:need2give/screens/main_pages_navbar/settings_page.dart';
 import 'package:need2give/screens/user/explore.dart';
-import 'package:need2give/screens/user/profile.dart';
 
 class ButtonNavbar extends StatefulWidget {
   static const String routeName = '/main';
@@ -20,7 +20,7 @@ class _ButtonNavbarState extends State<ButtonNavbar> {
   final List<Widget> _children = [
     const Dashboard(),
     const Explore(),
-    const Profile(),
+    const DonationCenterProfile(),
     const SettingsPage(),
   ];
