+++ need2give/lib/screens/user/donation_center/about_us.dart
@@ -12,7 +12,7 @@ class AboutUs extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: const EdgeInsets.all(8),
+      padding: const EdgeInsets.all(18),
       child: ListView(
         children: [
           const Text(
